+++ src/FerriteViz.jl
@@ -1,7 +1,6 @@
 module FerriteViz
 
 using Makie
-using StaticArrays
 using Tensors
 using Ferrite
 import Ferrite: get_grid, getrefshape
@@ -10,13 +9,26 @@ import ShaderAbstractions
 import LinearAlgebra
 
 abstract type AbstractPlotter end
+abstract type AbstractFilter end
 
-include("utils.jl")
-include("makieplotting.jl")
-include("lor_tools.jl")
+include("tessellation.jl")
+include("qptessellation.jl")
+include("dataset.jl")
+include("gradient.jl")
+include("filters.jl")
+include("representations.jl")
+include("viewer.jl")
 
-export MakiePlotter
-export ferriteviewer
-export for_discretization
+export FEData, apply
+export set_point_data!, set_cell_data!
+export WarpByVector, Gradient, CrinkleClip, Refine, FirstOrderRefinement, AddQuadraturePointData
+export ExtractComponent, Magnitude, Norm1, VonMises, Deviator, Threshold, Derive
+export ClipPlane
+export solutionplot, solutionplot!, cellplot, cellplot!, meshplot, meshplot!,
+       surfaceplot, surfaceplot!, arrowplot, arrowplot!, elementinfo, elementinfo!
+# composable viewer (SpecApi)
+export ferriteviewer, Control, default_controls, default_layout, default_pipeline
+export FieldMenu, ProcessMenu, ColormapMenu, WireframeToggle, LabelsToggle, DeformationToggle, TimeSlider
+export panelspec, solutionplotspec, meshplotspec, cellplotspec, surfaceplotspec, arrowplotspec
 
 end